chore(assets): consolidate peanut mascot assets into one canonical home
The mascot art was scattered and awkward: 4 GIFs lived in a separate
`peanut-animations` git submodule behind a `@/animations` alias with
cryptic names (512X512_ALPHA_GIF_konradurban_0N.gif), while stills sat in
src/assets/peanut. ~30 call sites imported raw submodule paths under
ad-hoc aliases (the same GIF was chillPeanutAnim / peanutAnim / PeanutGuyGIF).

Vendor everything into src/assets/mascot/ with emotion-named files and a
single labelled barrel; drop the submodule (+ .gitmodules entry, tsconfig
alias) and route every consumer through @/assets/mascot. Names now say what
they are: PeanutPeace / PeanutPointing / PeanutCheering / PeanutCrying /
PeanutThinking. Also retires the legless peanut-raising-hands.svg on the card
waitlist screen in favour of the cheering mascot.

No runtime behaviour change. Optimization of the GIFs is handed off separately.
