+++ src/app/maintenance/page.tsx
@@ -1,5 +1,5 @@
 'use client'
-import chillPeanutAnim from '@/animations/GIF_ALPHA_BACKGORUND/512X512_ALPHA_GIF_konradurban_01.gif'
+import { PeanutPeace } from '@/assets/mascot'
 import { Button } from '@/components/0_Bruddle/Button'
 import Image from 'next/image'
 import { useRouter } from 'next/navigation'
@@ -8,7 +8,7 @@ const MaintenancePage = () => {
     const router = useRouter()
     return (
         <div className="flex h-screen flex-col items-center justify-center gap-4 p-6">
-            <Image src={chillPeanutAnim.src} alt="Maintenance" width={250} height={250} />
+            <Image src={PeanutPeace.src} alt="Maintenance" width={250} height={250} />
             <h1 className="text-3xl font-bold text-black">Under Maintenance</h1>
             <p className="text-center text-lg text-gray-1">
                 We are currently going through maintenance. We should be back online shortly. Sorry for the
