+++ src/assets/illustrations/index.ts
@@ -1,4 +1,3 @@
-export { default as PeanutGuyGIF } from '@/animations/GIF_ALPHA_BACKGORUND/512X512_ALPHA_GIF_konradurban_01.gif'
 export { default as ButterySmoothGlobalMoney } from './buttery_smooth_global_money.svg'
 export { default as ClaimChainsBadge } from './claim-chains-badge.svg'
 export { default as Cloud } from './cloud.svg'
@@ -12,6 +11,5 @@ export { default as PEANUT_LOGO_BLACK } from './peanut-logo-dark.svg'
 export { default as PeanutsBG } from './peanuts-bg.svg'
 export { default as Sparkle } from './sparkle.svg'
 export { default as Star } from './star.svg'
-export { default as ThinkingPeanut } from './thinking_peanut.gif'
 export { default as LandingCountries } from './landing-countries.svg'
 export { default as GlobalCashLocalFeel } from './global-cash-local-feel.png'
