+++ src/assets/index.ts
@@ -6,5 +6,5 @@ export * from './exchanges'
 export * from './icons'
 export * from './illustrations'
 export * from './logos'
+export * from './mascot'
 export * from './payment-apps'
-export * from './peanut'
