+++ src/assets/mascot/index.ts
@@ -1,0 +1,17 @@
+// Canonical home for the Peanut mascot character assets.
+// Everything mascot-shaped lives here — animated GIFs and stills alike.
+// Import from '@/assets/mascot' only; do not reach for raw file paths.
+
+// Animated mascots (GIF, alpha background)
+export { default as PeanutPeace } from './peanut-peace.gif' // double peace-sign, mid-stride — the all-purpose "peanut guy"
+export { default as PeanutPointing } from './peanut-pointing.gif' // grinning, pointing off-screen
+export { default as PeanutCheering } from './peanut-cheering.gif' // both fists up, celebrating — success / "you're in"
+export { default as PeanutCrying } from './peanut-crying.gif' // teary, hands to face — errors / empty states
+export { default as PeanutThinking } from './peanut-thinking.gif' // pondering — loading / verification waits
+
+// Stills
+export { default as PEANUTMAN_LOGO } from './peanut-logo.svg'
+export { default as PEANUTMAN_WAVING } from './peanut-waving.svg'
+export { default as PEANUTMAN_PFP } from './peanut-pfp.svg'
+export { default as PEANUTMAN_HOLDING_BEER } from './peanut-beer.svg'
+export { default as PEANUTMAN_MOBILE } from './peanut-club.png'
