+++ src/assets/mascot/peanut-club.png
