+++ src/assets/mascot/peanut-thinking.gif
