--- src/assets/peanut/index.ts
@@ -1,6 +1,0 @@
-export { default as PEANUTMAN_MOBILE } from './peanut-club.png'
-export { default as PEANUTMAN_PFP } from './peanut-pfp.svg'
-export { default as PEANUTMAN_RAISING_HANDS } from './peanut-raising-hands.svg'
-export { default as PEANUTMAN_HOLDING_BEER } from './peanut-holding-beer.svg'
-export { default as PEANUTMAN_LOGO } from './peanutman-logo.svg'
-export { default as PEANUTMAN_WAVING } from './peanutman-waving.svg'
