+++ src/components/Claim/Generic/ClaimError.view.tsx
@@ -4,7 +4,7 @@ import { Button } from '@/components/0_Bruddle/Button'
 import { useModalsContext } from '@/context/ModalsContext'
 import Image from 'next/image'
 import Link from 'next/link'
-import PEANUTMAN_CRY from '@/animations/GIF_ALPHA_BACKGORUND/512X512_ALPHA_GIF_konradurban_05.gif'
+import { PeanutCrying } from '@/assets/mascot'
 
 type ClaimErrorViewProps = {
     title: string
@@ -18,7 +18,7 @@ export const ClaimErrorView = ({ title, message, primaryButtonText, onPrimaryCli
 
     return (
         <div className="flex flex-col items-center justify-center space-y-4 rounded-lg text-center">
-            <Image src={PEANUTMAN_CRY.src} alt="Sad peanut 😢" width={96} height={96} />
+            <Image src={PeanutCrying.src} alt="Sad peanut 😢" width={96} height={96} />
             <div className="space-y-2">
                 <h1 className="text-lg font-semibold">{title}</h1>
                 <p className="text-sm font-normal md:max-w-xs">{message}</p>
