+++ src/features/payments/shared/components/PaymentSuccessView.tsx
@@ -36,7 +36,7 @@ import Image from 'next/image'
 import { useRouter } from 'next/navigation'
 import { type ReactNode, useEffect, useMemo, useRef } from 'react'
 import { usePointsConfetti } from '@/hooks/usePointsConfetti'
-import chillPeanutAnim from '@/animations/GIF_ALPHA_BACKGORUND/512X512_ALPHA_GIF_konradurban_01.gif'
+import { PeanutPeace } from '@/assets/mascot'
 import { useHaptic } from 'use-haptic'
 import posthog from 'posthog-js'
 import { ANALYTICS_EVENTS } from '@/constants/analytics.consts'
@@ -260,7 +260,7 @@ const PaymentSuccessView = ({
             )}
             <div className="relative z-10 my-auto flex h-full flex-col justify-center space-y-4">
                 <Image
-                    src={chillPeanutAnim.src}
+                    src={PeanutPeace.src}
                     alt="Peanut Mascot"
                     width={20}
                     height={20}
